extension/bridge/requirements.txt
# Python Bridge Dependencies
# For TypeScript-Python subprocess communication with Cognee

# Core Cognee dependency
cognee==0.3.4

# Environment variable management
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
